lrs: dev_picker: forbid multiple parallel put to same group
The lrs scheduler would skip over busy drives and allow picking a new
tape if a mover has multiple drives available and multiple requests for
a single group are queued up.

This patch changes the priority to limit the number of tapes used for a
group. This is a tradeoff: if many requests for a single group come in
the throughput for this group will be lower, but if there only had been
a few requests the overhead of unused spaces in tapes will be lower.

The idea would be to maximize overall throughput through scheduling
multiple groups in parallel, instead of maximizing the throughput for
a single group.

Change-Id: I5b072e9bc87abe9a2f2f4888a3b56f0454e37eec